enhancement: address CodeRabbit review findings
- Extract duplicated resolveToken() into ResolvesServiceTokens trait
- Add pagination support to GitLabWikiService::getWikiPages()
- Prevent double-encoding of $ref in GitLabWikiService::getFileContent()
- Update test exception messages to match full production messages
